dxfeed streamer removed; cert/real sessions combined (#140)
* remove dxfeed streamer; new streamer URL; use same class for prod/cert sessions

* update docs

* update readme to handle github breakage

* use compact data format for streamer

* relax coverage, add 2fa

* fix isort

* fix candle bug

+++ tastytrade/dxfeed/candle.py
@@ -33,13 +33,13 @@ class Candle(Event):
     #: the last (close) price of the candle
     close: Optional[Decimal] = None
     #: the total volume of the candle
-    volume: Optional[int] = None
+    volume: Optional[Decimal] = None
     #: volume-weighted average price
     vwap: Optional[Decimal] = None
     #: bid volume in the candle
-    bidVolume: Optional[int] = None
+    bidVolume: Optional[Decimal] = None
     #: ask volume in the candle
-    askVolume: Optional[int] = None
+    askVolume: Optional[Decimal] = None
     #: implied volatility in the candle
     impVolatility: Optional[Decimal] = None
     #: open interest in the candle
